// src/app/models/admin/capphephoatdongkhoangsan/cptanthukhoangsan/cptanthukhuvuc.model.ts
export class OutputCpTanThuKhuVucModel {
  idtanthukhuvuc: string;
  tenkhuvuc: string;
  dientich: number;
  donvidientich: string;
  idcappheptanthu: string;
  loaikhuvuc: number;
  geom: string;
  geovn: string;
  geowgs: string;
  hequychieu: string;
  siteid: string;
}

export class InputCpTanThuKhuVucModel {
  idtanthukhuvuc: string;
  tenkhuvuc: string;
  dientich: number;
  donvidientich: string;
  idcappheptanthu: string;
  loaikhuvuc: number;
  hequychieu: string;
  loaicapphep: string;
  listtoado = [];
}